refactor: split GitHubAdapter into adapter-github.ts to stay under 300 LOC
adapter.ts (115 LOC): types, OrchestratorAdapter interface, factory
adapter-github.ts (252 LOC): GitHubAdapter implementation

Addresses Gate 4 code smell finding — adapter.ts was 351 lines.
All existing exports remain available via adapter.ts re-exports.


Aloop-Agent: build
Aloop-Iteration: 30
Aloop-Session: orchestrator-20260321-172932-issue-176-20260327-063441
