+++ TODO.md
@@ -11,14 +11,9 @@
 
 ### In Progress
 
-- [x] [review] Gate 4: Dead import in `orchestrate.ts:19` — `import { createAdapter, type OrchestratorAdapter }` is present but neither symbol is used anywhere in the file (grep returns only the import line). Constitution Rule 13: no dead code. Remove the unused import. (priority: high)
-
-- [x] [qa/P1] index.test.ts "index CLI catches errors and prints clean messages without stack traces" fails consistently: test runs `aloop orchestrate --autonomy-level invalid` expecting `^Error: Invalid autonomy level: invalid` on stderr, but receives `ERR_MODULE_NOT_FOUND`. Pre-existing before this branch. 1/5 index tests fail. (priority: high)
-
-- [ ] [review] Gate 4: `adapter.ts` is 351 lines — above the 300 LOC "code smell" threshold from SPEC-ADDENDUM.md ("Files above 300 LOC are a code smell and must be decomposed before adding new features"). The four new methods (`closePr`, `getPrDiff`, `queryPrs`, `checkBranchExists`) added ~56 lines and pushed the file over the limit. Split the GitHubAdapter implementation methods into a separate file (e.g., `adapter-github.ts`) and keep only the interface types + factory in `adapter.ts`, or group and export from sub-files. (priority: medium)
-
 ### Completed
 
+- [x] [review] Gate 4: Split `GitHubAdapter` into `adapter-github.ts` (252 LOC) — `adapter.ts` now 115 LOC, both under 300 LOC threshold. (priority: medium)
 - [x] `OrchestratorAdapter` interface defined in `src/lib/adapter.ts` (lines 61–102)
 - [x] `GitHubAdapter` class implements full interface with GHE support
 - [x] `closePr`, `getPrDiff`, `queryPrs`, `checkBranchExists` added to interface and GitHubAdapter
@@ -28,3 +23,5 @@
 - [x] GHE URL support via `ghHost` config and `GH_HOST` env var
 - [x] No hardcoded `github.com` API URLs in built artifact
 - [x] LocalAdapter descoped from this PR (tracked separately)
+- [x] [review] Gate 4: Dead import in `orchestrate.ts:19` — `import { createAdapter, type OrchestratorAdapter }` is present but neither symbol is used anywhere in the file (grep returns only the import line). Constitution Rule 13: no dead code. Remove the unused import. (priority: high)
+- [x] [qa/P1] index.test.ts "index CLI catches errors and prints clean messages without stack traces" fails consistently: test runs `aloop orchestrate --autonomy-level invalid` expecting `^Error: Invalid autonomy level: invalid` on stderr, but receives `ERR_MODULE_NOT_FOUND`. Pre-existing before this branch. 1/5 index tests fail. (priority: high)
